requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.8.2
pypdf==4.3.1
python-docx==1.1.2
openai==1.44.0
pinecone-client==5.0.1
httpx==0.27.2  
